Refactor common logic into utils
