+++ utils/parseWeight.js
@@ -1,0 +1,3 @@
+export default function parseWeight(wt = 0) {
+  return parseInt(wt) || 0;
+}
